% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{Someonse's name I like}
}
\description{
hello
}
\examples{
hello(name = "Martin")
}
\author{
MMag. Martin Cohen MSc (WU), BSc(WU)
}
